pyproject.toml
[project]
name = "airtrail-flightera"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
